requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
idna==3.10
python-dotenv==1.1.0
requests==2.32.3
typing==3.7.4.3
urllib3==2.3.0
validators==0.34.0
